m3-auto-reporter/requirements.txt
pandas~=2.2
python-dotenv~=1.0
google-cloud-bigquery~=3.25
google-cloud-aiplatform~=1.101
requests~=2.32
